// src/lib/cardcom/generated/models/documentToCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type DocumentToCreate = typeof DocumentToCreate[keyof typeof DocumentToCreate];


export const DocumentToCreate = {
  Auto: 'Auto',
  TaxInvoiceAndReceipt: 'TaxInvoiceAndReceipt',
  TaxInvoiceAndReceiptRefund: 'TaxInvoiceAndReceiptRefund',
  Receipt: 'Receipt',
  ReceiptRefund: 'ReceiptRefund',
  Quote: 'Quote',
  Order: 'Order',
  OrderConfirmation: 'OrderConfirmation',
  OrderConfirmationRefund: 'OrderConfirmationRefund',
  DeliveryNote: 'DeliveryNote',
  DeliveryNoteRefund: 'DeliveryNoteRefund',
  ProformaInvoice: 'ProformaInvoice',
  DemandForPayment: 'DemandForPayment',
  DemandForPaymentRefund: 'DemandForPaymentRefund',
  ProformaDealInvoice: 'ProformaDealInvoice',
  ProformaDealInvoiceRefund: 'ProformaDealInvoiceRefund',
  TaxInvoice: 'TaxInvoice',
  ProformaInvoiceRefund: 'ProformaInvoiceRefund',
  TaxInvoiceRefund: 'TaxInvoiceRefund',
  ReceiptForTaxInvoice: 'ReceiptForTaxInvoice',
  DonationReceipt: 'DonationReceipt',
  DonationReceiptRefund: 'DonationReceiptRefund',
  CouponDocumentAndReceipt: 'CouponDocumentAndReceipt',
  CouponDocumentAndReceiptRefund: 'CouponDocumentAndReceiptRefund',
  ReceiptForTaxInvoiceRefund: 'ReceiptForTaxInvoiceRefund',
} as const;