// src/lib/cardcom/generated/models/operation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */

export type Operation = typeof Operation[keyof typeof Operation];


export const Operation = {
  ChargeOnly: 'ChargeOnly',
  ChargeAndCreateToken: 'ChargeAndCreateToken',
  CreateTokenOnly: 'CreateTokenOnly',
  SuspendedDeal: 'SuspendedDeal',
  Do3DSAndSubmit: 'Do3DSAndSubmit',
} as const;