// src/lib/cardcom/generated/models/createLowProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CardCom API
 * OpenAPI spec version: 11.0
 */
import type { AdvancedLPDefinition } from './advancedLPDefinition';
import type { DocumentLP } from './documentLP';
import type { LPUtmData } from './lPUtmData';
import type { Operation } from './operation';
import type { UIDefinition } from './uIDefinition';

/**
 * create a new one time, Iframe \ redirect page
 */
export interface CreateLowProfile {
  /** Cardcom terminal number */
  TerminalNumber: number;
  /**
     * Api Name for authentication
     * @minLength 1
     */
  ApiName: string;
  /** Type of operation to perform , default - ChargeOnly */
  Operation?: Operation;
  /**
     * A string of data to save on the transaction, usually send your unique order Id, you will get it back in the WebHook URL
     * @maxLength 250
     * @nullable
     */
  ReturnValue?: string | null;
  /** Amount of tranzaction (12.36) */
  Amount: number;
  /**
     * Url to redirect the card holder on successful transaction , (http(s)://site.com/Success)
     * @minLength 1
     * @maxLength 500
     */
  SuccessRedirectUrl: string;
  /**
     * Url to redicet on failed transaction (http(s)://site.com/failed)
     * @minLength 1
     * @maxLength 500
     */
  FailedRedirectUrl: string;
  /**
     * Url to redicet on cancel button Pressed (http(s)://site.com/failed)
     * @maxLength 500
     * @nullable
     */
  CancelRedirectUrl?: string | null;
  /**
     * A webhook URL is a web address that receives real-time data about the transaction
     * @minLength 1
     * @maxLength 500
     */
  WebHookUrl: string;
  /**
     * Text to show to card holder if no documents is send
     * @maxLength 250
     * @nullable
     */
  ProductName?: string | null;
  /**
     * Page Language - he,en,ru,sp (https://cardcomapinametovalue.zendesk.com/hc/he/articles/27007768777234-שפות-לדף-פרופיל-נמוך-Language)
     * @maxLength 2
     * @nullable
     */
  Language?: string | null;
  /**
     * ISO Coin Code ,1 - ILS, 2 - USD, else ISO Coin from list : https://en.wikipedia.org/wiki/ISO_4217
     * @nullable
     */
  ISOCoinId?: number | null;
  /** UI Definition for the created page */
  UIDefinition?: UIDefinition | null;
  /** Advanced transaction definition */
  AdvancedDefinition?: AdvancedLPDefinition | null;
  /** Document information */
  Document?: DocumentLP | null;
  /** Document information */
  UTM?: LPUtmData | null;
}